refactor(infra-messaging): align option-token internalization strategy
