+++ packages/cron/README.md
@@ -116,8 +116,8 @@ Timeout behavior: after a timeout task fires, its task definition remains in the
 ### Root barrel public surface governance (0.x)
 
 - **supported**: scheduling decorators (`@Cron`, `@Interval`, `@Timeout`), `CronExpression`, `CronModule.forRoot`, `createCronProviders`, `SCHEDULING_REGISTRY`, and status snapshot helpers.
-- **compatibility-only**: `CRON_OPTIONS` is kept as an internal compatibility token (no longer part of root exports). `normalizeCronModuleOptions` and metadata helper exports (`defineSchedulingTaskMetadata`, `defineCronTaskMetadata`, `get*TaskMetadata*`, `schedulingMetadataSymbol`, `cronMetadataSymbol`) remain exported for 0.x compatibility and framework/tooling integration, but are not recommended for new app-level imports.
-- **internal**: scheduler lifecycle internals beyond documented APIs are not part of the root-barrel contract.
+- **compatibility-only**: `normalizeCronModuleOptions` and metadata helper exports (`defineSchedulingTaskMetadata`, `defineCronTaskMetadata`, `get*TaskMetadata*`, `schedulingMetadataSymbol`, `cronMetadataSymbol`) remain exported for 0.x compatibility and framework/tooling integration, but are not recommended for new app-level imports.
+- **internal**: `CRON_OPTIONS` and scheduler lifecycle internals beyond documented APIs are not part of the root-barrel contract.
 
 ## non-goals and intentional limitations
