--- packages/queue/src/internal-tokens.ts
@@ -1,5 +1,0 @@
-import type { Token } from '@konekti/core';
-
-import type { NormalizedQueueModuleOptions } from './types.js';
-
-export const QUEUE_OPTIONS: Token<NormalizedQueueModuleOptions> = Symbol.for('konekti.queue.options');
